backend/pyproject.toml
[tool.poetry]
name = "storai-booker-backend"
version = "0.1.0"
description = "StorAI-Booker Backend - AI-powered storybook and comic book generation"
authors = ["StorAI Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
motor = "^3.3.2"
beanie = "^1.24.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
boto3 = "^1.34.0"
pillow = "^10.2.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"
langchain-google-genai = "^2.0.0"
google-generativeai = "^0.8.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
slowapi = "^0.1.9"
httpx = "^0.26.0"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
ruff = "^0.1.11"
black = "^23.12.1"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
